Fix type cast error for 64bit compilation
RSP_LoadMatrixX86.cpp:92:28: error: cast from 'u8* {aka unsigned char*}' to 'int' loses precision [-fpermissive]

+++ src/RSP_LoadMatrixX86.cpp
@@ -89,7 +89,7 @@ LoadLoop:
     "    loop    LoadLoop"                       "\n\t"
     ".att_syntax prefix"                         "\n\t"
     : /* no output */
-    : "f"(recip), "S"((int)RDRAM+address), "D"(mtx), "c"(4)
+    : "f"(recip), "S"((intptr_t)RDRAM+address), "D"(mtx), "c"(4)
     : "memory" );
 #endif // WIN32_ASM
 }
